Fix: reservation details page (#58)
* fix: rename ReservationApproval to ReservationDetails

* feat: dynamically display cancellation and submit buttons

* fix: style issue & link api page

* fix: redirect after submit on personal data page
