+++ src/components/FloatingSidebar.tsx
@@ -15,7 +15,7 @@ export default function FloatingSidebar() {
         </IconButton>
       </Tooltip>
       <Tooltip title='Documentation' placement='right' arrow>
-        <IconButton>
+        <IconButton href='https://reservation-bear.de/api/' target='_blank'>
           <BookIcon fontSize='large' className='social-icon' />
         </IconButton>
       </Tooltip>
