--- src/pages/RestaurantDetails.tsx
@@ -1,7 +1,0 @@
-/**
- * Restaurant Detail Page
- * @return {JSX.Element}
- */
-export default function RestaurantDetails() {
-  return <div>RestaurantDetails</div>;
-}
